Move ReactionRecord from common to message becoz it is used only there
